+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this branch will be documented in this file.
 
 - Added RTL (Right-to-Left) support for Arabic like languages, allowing the app to display correctly in these languages.
 - Introduced a `ContentDialog` to display scan results on the `Repair` page #44.
+- Added support for Hebrew language (`he-il`), contributed by @Y-PLONI #36.
 
 ### Fixes
 
@@ -16,7 +17,6 @@ All notable changes to this branch will be documented in this file.
 
 ### Changes
 
-- Didn't add the Hebrew language support #36.
 - Changed the restore point creation process to use the default System Restore `SystemPropertiesProtection` #41.
 - Enhance telemetry disabling and temp file cleanup.